Migrate from html-pdf to puppeteer
Sadly with the migration to Puppeteer we're going to make headers and
footers significantly worse due to a number of bugs and incomplete
features in Puppeteers headers and footers. In an effort to move mdpdf
forward we'll make a release with these changes and see what people
say. After all, the best way to get feedback on the internet is to be wrong.